fix my_put_nbr

+++ lib/my/fonctions/get_nbrlen.c
@@ -5,7 +5,7 @@
 ** get_nbrlen
 */
 
-unsigned int get_nbrlen(int nb)
+unsigned int get_nbrlen(long long nb)
 {
     return ((nb != 0) ?
     (1 + get_nbrlen(nb / 10)) : 0);
